back-end/fastapi/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
gunicorn = "^21.2.0"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
aiofiles = "^23.2.1"


[tool.poetry.group.data-science.dependencies]
numpy = "^1.26.2"


[tool.poetry.group.llm.dependencies]
langchain = "^0.1.14"
llama-index = "^0.10.26"
qdrant-client = "^1.8.2"
mecab-python3 = "^1.0.8"
openai = "^1.14.3"
langchain-community = "^0.0.30"
langchain-openai = "^0.1.1"


[tool.poetry.group.document.dependencies]
ndjson = "^0.3.1"
pymupdf = "^1.23.6"


[tool.poetry.group.database.dependencies]
sqlalchemy = "^2.0.23"
aiomysql = "^0.2.0"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.2"
aiosqlite = "^0.19.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
